Restructure project for best practices: organize into package, examples, and docs directories
