+++ README.md
@@ -210,13 +210,22 @@ python3 ppt_assistant.py example_config.json
 
 ```
 ppt_assistant/
-├── ppt_assistant.py      # 主程序
+├── ppt_assistant/        # 核心包
+│   ├── __init__.py       # 包初始化
+│   └── core.py           # 核心功能實現
+├── ppt_assistant.py      # 主入口點
 ├── README.md             # 使用文檔
-├── json_format.md        # JSON 格式詳細說明
-├── example_config.json   # 示例配置文件
-├── template.pptx         # 示例模板文件 (16:9)
-├── sample_image.png      # 示例圖片
-└── output_presentation.pptx  # 生成的示例簡報 (16:9)
+├── examples/             # 示例文件
+│   ├── example_config.json   # 示例配置文件
+│   ├── template.pptx         # 示例模板文件 (16:9)
+│   ├── sample_image.png      # 示例圖片
+│   └── output_presentation.pptx  # 生成的示例簡報 (16:9)
+└── docs/                 # 文檔
+    ├── FORMATTING_GUIDE.md   # 格式化指南
+    ├── json_format.md        # JSON 格式詳細說明
+    ├── LAYOUT_SPECS.md       # 佈局規範
+    ├── PROJECT_SUMMARY.md    # 項目總結
+    └── QUICKSTART.md         # 快速開始
 ```
 
 ## 進階使用
