+++ requirements.txt
@@ -1,0 +1,1 @@
+python-pptx>=0.6.21
